+++ src/App.js
@@ -1,18 +1,11 @@
 import "bootstrap/dist/css/bootstrap.min.css";
 
 import Dashboard from './components/Dashboard';
+import DashboardOverview from "./components/DashboardOverview/DashboardOverview";
 
 import styles from './App.module.scss'
 
 function App() {
-
-  const handleChange = () => {
-    console.log('6== change');
-  }
-
-  const handleBlur = () => {
-    console.log('9== blur');
-  }
 
   return (
     <div className={`${styles.app} container-fluid`}>
@@ -21,13 +14,7 @@ function App() {
           <Dashboard />
         </section>
         <section className={`${styles.content_container} col-sm-10`}>
-          <header className={`d-flex justify-content-between`}>
-            <div>
-              <h6> Hello Shathruk</h6></div>
-            <div>
-              <input type='text' placeholder='Search' onChange={handleChange} onBlur={handleBlur} />
-            </div>
-          </header>
+          <DashboardOverview />
         </section>
       </div>
     </div>
