+++ src/App.module.scss
@@ -1,14 +1,13 @@
 .app {
-  // text-align: center;
-
   .navigation_container {
     background-color: #040440;
     height: 100vh;
     color: #ffffff;
   }
+
   .content_container {
-background-color: f5f6f8;
-color: #000000;
+    background-color: #f5f6f8;
+    color: #000000;
+    padding: 20px 20px 10px;
   }
 }
-
