+++ src/components/DashboardOverview/DashboardOverview.module.scss
@@ -1,0 +1,17 @@
+.search_container {
+    position: relative;
+
+    >i {
+        position: absolute;
+        left: 5px;
+        top: 5px;
+    }
+
+    >input {
+        border: 1px solid white;
+        border-radius: 4px;
+        padding-left: 24px;
+        padding-top: 3px;
+        padding-bottom: 3px;
+    }
+}
